+++ apps/playground/src/components/TheHeader.vue
@@ -62,14 +62,14 @@ const copyLink = async () => {
         target="_blank"
         aria-label="View on GitHub"
       >
-        <HeaderIconButton label="View on GitHub" :icon="githubLogo" variation="secondary" />
+        <HeaderIconButton label="View on GitHub" :icon="githubLogo" />
       </a>
     </div>
   </header>
 </template>
 
 <style lang="scss" scoped>
-@use "sit-onyx/breakpoints.scss" as onyx;
+@use "sit-onyx/breakpoints.scss";
 
 .header {
   box-sizing: border-box;
@@ -106,7 +106,7 @@ const copyLink = async () => {
     width: 10rem;
   }
 
-  @include onyx.breakpoint(max, md) {
+  @include breakpoints.screen(max, md) {
     flex-direction: column;
     align-items: flex-start;