+++ packages/sit-onyx/src/types/index.ts
@@ -1,3 +1,4 @@
+export * from "./breakpoints";
 export * from "./colors";
 export * from "./fonts";
 export * from "./forms";