+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,90 @@
 # alpha-test-app
 
+## 0.1.0-alpha.102
+
+### Patch Changes
+
+- Updated dependencies [c011e27]
+  - sit-onyx@1.0.0-alpha.105
+
+## 0.1.0-alpha.101
+
+### Patch Changes
+
+- Updated dependencies [20fb878]
+  - sit-onyx@1.0.0-alpha.104
+
+## 0.1.0-alpha.100
+
+### Patch Changes
+
+- Updated dependencies [09ce727]
+  - sit-onyx@1.0.0-alpha.103
+
+## 0.1.0-alpha.99
+
+### Patch Changes
+
+- Updated dependencies [3b9fbf4]
+  - sit-onyx@1.0.0-alpha.102
+
+## 0.1.0-alpha.98
+
+### Patch Changes
+
+- Updated dependencies [2ea735b]
+  - sit-onyx@1.0.0-alpha.101
+
+## 0.1.0-alpha.97
+
+### Patch Changes
+
+- Updated dependencies [d54d357]
+  - sit-onyx@1.0.0-alpha.100
+
+## 0.1.0-alpha.96
+
+### Patch Changes
+
+- Updated dependencies [3a9783d]
+  - sit-onyx@1.0.0-alpha.99
+
+## 0.1.0-alpha.95
+
+### Patch Changes
+
+- Updated dependencies [9eb7b4e]
+- Updated dependencies [9eb7b4e]
+  - sit-onyx@1.0.0-alpha.98
+
+## 0.1.0-alpha.94
+
+### Patch Changes
+
+- Updated dependencies [5754525]
+  - sit-onyx@1.0.0-alpha.97
+
+## 0.1.0-alpha.93
+
+### Patch Changes
+
+- Updated dependencies [f99f38f]
+  - sit-onyx@1.0.0-alpha.96
+
+## 0.1.0-alpha.92
+
+### Patch Changes
+
+- Updated dependencies [2f825ec]
+  - sit-onyx@1.0.0-alpha.95
+
+## 0.1.0-alpha.91
+
+### Patch Changes
+
+- Updated dependencies [c867746]
+  - sit-onyx@1.0.0-alpha.94
+
 ## 0.1.0-alpha.90
 
 ### Patch Changes