+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,90 @@
 # playground
 
+## 0.0.1-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [c011e27]
+  - sit-onyx@1.0.0-alpha.105
+
+## 0.0.1-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [20fb878]
+  - sit-onyx@1.0.0-alpha.104
+
+## 0.0.1-alpha.42
+
+### Patch Changes
+
+- Updated dependencies [09ce727]
+  - sit-onyx@1.0.0-alpha.103
+
+## 0.0.1-alpha.41
+
+### Patch Changes
+
+- Updated dependencies [3b9fbf4]
+  - sit-onyx@1.0.0-alpha.102
+
+## 0.0.1-alpha.40
+
+### Patch Changes
+
+- Updated dependencies [2ea735b]
+  - sit-onyx@1.0.0-alpha.101
+
+## 0.0.1-alpha.39
+
+### Patch Changes
+
+- Updated dependencies [d54d357]
+  - sit-onyx@1.0.0-alpha.100
+
+## 0.0.1-alpha.38
+
+### Patch Changes
+
+- Updated dependencies [3a9783d]
+  - sit-onyx@1.0.0-alpha.99
+
+## 0.0.1-alpha.37
+
+### Patch Changes
+
+- Updated dependencies [9eb7b4e]
+- Updated dependencies [9eb7b4e]
+  - sit-onyx@1.0.0-alpha.98
+
+## 0.0.1-alpha.36
+
+### Patch Changes
+
+- Updated dependencies [5754525]
+  - sit-onyx@1.0.0-alpha.97
+
+## 0.0.1-alpha.35
+
+### Patch Changes
+
+- Updated dependencies [f99f38f]
+  - sit-onyx@1.0.0-alpha.96
+
+## 0.0.1-alpha.34
+
+### Patch Changes
+
+- Updated dependencies [2f825ec]
+  - sit-onyx@1.0.0-alpha.95
+
+## 0.0.1-alpha.33
+
+### Patch Changes
+
+- Updated dependencies [c867746]
+  - sit-onyx@1.0.0-alpha.94
+
 ## 0.0.1-alpha.32
 
 ### Patch Changes