+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,106 @@
 # sit-onyx
 
+## 1.0.0-alpha.105
+
+### Patch Changes
+
+- c011e27: fix(grid): include padding in max-width
+
+## 1.0.0-alpha.104
+
+### Minor Changes
+
+- 20fb878: feat(theme): include color-scheme in the onyx dark/light theme
+
+## 1.0.0-alpha.103
+
+### Patch Changes
+
+- 09ce727: fix(OnyxCheckbox): align checkbox label padding according to density, stabilize check all border
+
+## 1.0.0-alpha.102
+
+### Minor Changes
+
+- 3b9fbf4: feat: add `OnyxTextarea` component
+
+## 1.0.0-alpha.101
+
+### Patch Changes
+
+- 2ea735b: Hide nested children when nav item is closed
+
+## 1.0.0-alpha.100
+
+### Patch Changes
+
+- d54d357: fix(OnyxUserMenu): hide flyout menu when closed
+
+## 1.0.0-alpha.99
+
+### Patch Changes
+
+- 3a9783d: Fix listbox position inside OnyxNavItem
+
+## 1.0.0-alpha.98
+
+### Major Changes
+
+- 9eb7b4e: rename SCSS breakpoint mixin
+
+  Old:
+
+  ```scss
+  @use "sit-onyx/breakpoints.scss" as onyx;
+
+  @include onyx.breakpoint(max, md) {
+    // your styles
+  }
+  ```
+
+  New:
+
+  ```scss
+  @use "sit-onyx/breakpoints.scss";
+
+  @include breakpoints.screen(max, md) {
+    // your styles
+  }
+  ```
+
+### Minor Changes
+
+- 9eb7b4e: feat: add `OnyxNavBar` component
+
+  If you used one of the `onyx-grid-max-md`, `onyx-grid-max-lg` or `onyx-grid-center` CSS classes which are not placed on the application root, move them to the application root element. See [grid docs](https://onyx.schwarz/development/grid.html#example) for further information
+
+## 1.0.0-alpha.97
+
+### Minor Changes
+
+- 5754525: feat(OnyxCheckboxGroup): align style for "check all" with figma
+
+## 1.0.0-alpha.96
+
+### Major Changes
+
+- f99f38f: refactor: align color property names
+
+  - rename property `variation` to `color` for: OnyxBadge, OnyxButton and OnyxIconButton
+  - OnyxButton and OnyxIconButton: rename color `secondary` to `neutral`
+
+## 1.0.0-alpha.95
+
+### Patch Changes
+
+- 2f825ec: fix(OnyxCheckboxGroup,OnyxListbox): preserve disabled states when toggling "check all"
+
+## 1.0.0-alpha.94
+
+### Patch Changes
+
+- c867746: fix(OnyxUserMenu): emit `optionClick` event when clicking an option
+
 ## 1.0.0-alpha.93
 
 ### Minor Changes